backend/requirements.txt
# c:\real_estate_project\backend\requirements.txt

# --- Flask Core Framework ---
Flask==3.0.3
Werkzeug==3.0.3

# --- Database (SQLAlchemy with SQLite) ---
Flask-SQLAlchemy==3.1.1

# --- Authentication (JWT) ---
Flask-JWT-Extended==4.6.0

# --- Data Science / Machine Learning ---
pandas==2.2.2
numpy==1.26.4          # CRITICAL: Added for log transformation
scikit-learn==1.4.2
joblib==1.4.2

# --- Ensemble Model (HIGH ACCURACY) ---
# We will use RandomForestRegressor from scikit-learn, so no new packages needed here, 
# but if you wanted XGBoost, you'd add:
# xgboost==2.0.3

# --- Email Functionality ---
Flask-Mail==0.9.1 

# --- Generative AI Chatbot ---
google-genai==0.8.0